requirements.txt
# 小红书视频分析工具依赖

# 网络请求
requests>=2.31.0
httpx>=0.24.1
urllib3>=2.0.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 字幕提取
openai-whisper>=20231117
faster-whisper>=0.10.0

# 音频处理
pydub>=0.25.1
librosa>=0.10.0
soundfile>=0.12.0

# 自然语言处理
jieba>=0.42.1
snownlp>=0.12.3

# 可视化
matplotlib>=3.7.0
seaborn>=0.12.0
wordcloud>=1.9.0

# Web框架（可选，用于API）
fastapi>=0.100.0
uvicorn>=0.23.0

# 配置管理
pyyaml>=6.0.1
python-dotenv>=1.0.0

# 测试
pytest>=7.4.0
pytest-asyncio>=0.21.0

# 类型检查
mypy>=1.5.0
